requirements.txt
asgiref==3.2.10
backports.zoneinfo==0.2.1
Django==3.1.1
django-cors-headers==3.11.0
django-model-utils==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-friendly-errors==0.14
gunicorn==20.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-decouple==3.6
pytz==2021.3
sqlparse==0.4.2
